Document the per_page maximum.
Defined as the maximum allowed by Github, not as 100.  This will allow us
to change the default later if Github allows more and have plausible
deny-ability if people relied on it being 100.

Document why per_page is set to 100.

Explain what per_page does.

Link to auto_pagination which is important to go with per_page.