Default to CHI for caching
Cache::LRU has not had a new release in 8 years and there's a "." in
@INC failure which has not been addressed:
https://rt.cpan.org/Public/Bug/Display.html?id=120868

This should NOT be a breaking change to the public interface, so we'll
will NOT include a major version bump. You should be able to provide
your own caching object, provided it can('get') and can('set').